Introducing content class
