+++ app/models/concerns/with_description.rb
@@ -1,0 +1,12 @@
+module WithDescription
+  extend ActiveSupport::Concern
+
+  included do
+    markdown_on :description, :description_teaser
+    validates_presence_of :name, :description
+  end
+
+  def description_teaser
+    description.markdown_paragraphs.first
+  end
+end
